% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BE_NIS_POSTAL}
\alias{BE_NIS_POSTAL}
\title{Link between postal code and NIS code at municipality level in Belgium.}
\source{
\url{https://vlaamseoverheid.atlassian.net/wiki/spaces/MG/pages/1042712614/Codetabellen}
}
\description{
Link between postal code and NIS code at municipality level in Belgium containing:
\itemize{
\item POSTAL_CODE: Postal code
\item NIS_CODE: NIS code
}
}
\examples{
data(BE_NIS_POSTAL)
str(BE_NIS_POSTAL)
}
\references{
\url{https://vlaamseoverheid.atlassian.net/wiki/spaces/MG/pages/1042712614/Codetabellen}
}
